anamoly_app/requirements.txt
paho-mqtt==1.5.0
numpy==1.19.5
scikit-learn==0.24.2
